requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
asyncpg
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib==1.7.4
pydantic>=2.0.0
pydantic-settings
python-multipart
bcrypt==4.0.1
databases
